Fix order of imports
